huggingface_space/requirements.txt
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
gradio>=4.0.0
spaces>=0.19.0
